requirements.txt
absl-py==1.4.0
altair==4.2.2
apturl==0.5.2
astunparse==1.6.3
attrs==22.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
filelock==3.12.2
Flask==2.2.3
flatbuffers==23.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.51.3
h5py==3.8.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
mysql-connector-python==8.0.32
netifaces==0.11.0
numpy==1.24.2
nvidia-ml-py==11.525.84
nvitop==1.0.0
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.8.1
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.0
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
validators==0.20.0
virtualenv==20.23.1
wadllib==1.3.6
watchdog==2.3.1
Werkzeug==2.2.3
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
